Bug: #2331
The web gui validation of filesystem watermarks was not working with decimal values such as 90.2 and 90.1 for high and low watermark.  The problem was tracked down to parsing as an int one of the watermark values, which caused the problem.  Its fixed now.

+++ ImageServer/Web/Common/WebControls/Validators/RangeComparisonValidator.js
@@ -44,7 +44,7 @@ function @@CLIENTID@@_ClientSideEvaluator()
             
         if (compareCtrl!=null && compareCtrl.value!='' &&  !isNaN(compareCtrl.value))
         {
-            compareValue = parseInt(compareCtrl.value);
+            compareValue = parseFloat(compareCtrl.value);
             result.OK = controlValue >= @@MIN_VALUE@@ && controlValue<= @@MAX_VALUE@@ && controlValue @@COMPARISON_OP@@ compareValue;
         }
         else
